[#211] feat: 임시 글 발행 시 포스트 이미지, 인트로 받지 않도록 변경
