+++ src/main/java/com/mallang/post/application/command/UpdateDraftCommand.java
@@ -9,9 +9,7 @@ public record UpdateDraftCommand(
         Long memberId,
         Long draftId,
         String title,
-        String intro,
         String bodyText,
-        @Nullable String postThumbnailImageName,
         @Nullable Long categoryId,
         List<String> tags
 ) {
