+++ src/main/java/com/mallang/post/presentation/request/CreateDraftRequest.java
@@ -7,9 +7,7 @@ import java.util.List;
 public record CreateDraftRequest(
         String blogName,
         String title,
-        @Nullable String intro,
         String bodyText,
-        @Nullable String postThumbnailImageName,
         @Nullable Long categoryId,
         List<String> tags
 ) {
@@ -19,8 +17,6 @@ public record CreateDraftRequest(
                 .blogName(blogName)
                 .title(title)
                 .bodyText(bodyText)
-                .postThumbnailImageName(postThumbnailImageName)
-                .intro(intro)
                 .categoryId(categoryId)
                 .tags(tags)
                 .build();
