+++ src/test/java/com/mallang/post/domain/PostTest.java
@@ -218,6 +218,36 @@ class PostTest {
             // when & then
             assertThat(post.getPostThumbnailImageName()).isNull();
         }
+
+        @Test
+        void 인트로가_없으면_본문에서_가져와_설정된다() {
+            // when
+            Post post = Post.builder()
+                    .blog(blog)
+                    .title("제목")
+                    .bodyText("내용")
+                    .writer(mallang)
+                    .build();
+
+            // then
+            String postIntro = post.getPostIntro();
+            assertThat(postIntro).isEqualTo("내용");
+        }
+
+        @Test
+        void 인트로가_없을때_본문이_150자_이상이면_150자만_인트로로_설정된다() {
+            // when
+            Post post = Post.builder()
+                    .blog(blog)
+                    .title("제목")
+                    .bodyText("1".repeat(151))
+                    .writer(mallang)
+                    .build();
+
+            // then
+            String postIntro = post.getPostIntro();
+            assertThat(postIntro.length()).isEqualTo(150);
+        }
     }
 
     @Nested
